Move getUser function from UserRegistrationServlet to UserService
